The following and followers actions. red
The only tricky part of the implementation is realizing that we need to
add two new actions to the Users controller. Based on the routes defined
in commit: Adding following and followers actions to the Users
controller, we need to call them following and followers. Each action
needs to set a title, find the user, retrieve either @user.following or
@user.followers (in paginated form), and then render the page. The
result appears in this commit.