The show_follow view used to render following and followers. green
As we’ve seen throughout this tutorial, the usual Rails convention is to
implicitly render the template corresponding to an action, such as
rendering show.html.erb at the end of the show action. In contrast, both
actions in commit: The following and followers actions make an explicit
call to render, in this case rendering a view called show_follow, which
we must create. The reason for the common view is that the ERb is nearly
identical for the two cases, and this commit that covers them both.

The actions in commit: The following and followers actions render the
view from commit: The following and followers actions in two contexts,
“following” and “followers”, with the results shown in [Figure
14.16](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/user_following_3rd_edition.png)
and [Figure
14.17](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/user_followers_3rd_edition.png).
Note that nothing in the above code uses the current user, so the same
links work for other users, as shown in [Figure
14.18](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/diferent_user_followers_3rd_edition.png).

At this point, the tests in this commit: Tests for the authorization of
the following and followers pages should be green due to the before
filter in commit: The following and followers actions:

```
$ rails test
```