Relationships fixtures for use in following/follower tests.
To test the show_follow rendering, we’ll write a couple of short
integration tests that verify the presence of working following and
followers pages. They are designed to be a sanity check, not to be
comprehensive; indeed, as noted in Section 5.3.4: Layout link tests,
comprehensive tests of things like HTML structure are likely to be
brittle and thus counter-productive. Our plan in the case of
following/followers pages is to check the number is correctly displayed
and that links with the right URLs appear on the page.

To get started, we’ll generate an integration test as usual:

```
$ rails generate integration_test following
      invoke  test_unit
      create    test/integration/following_test.rb
```

Next, we need to assemble some test data, which we can do by adding some
relationships fixtures to create following/follower relationships.
Recall from Section 13.2.3: Profile micropost tests that we can use code
like

```
orange:
  content: "I just ate an orange!"
  created_at: <%= 10.minutes.ago %>
  user: michael
```

to associate a micropost with a given user. In particular, we can write

```
user: michael
```
instead of

```
user_id: 1
```
Applying this idea to the relationships fixtures gives the associations
in this commit.