Tests for following/follower pages. green
The fixtures in commit: Relationships fixtures for use in
following/follower tests first arrange for Michael to follow Lana and
Malory, and then arrange for Michael to be followed by Lana and Archer.
To test for the right count, we can use the same assert_match method we
used in commit: A test for the user profile to test for the display of
the number of microposts on the user profile page. Adding in assertions
for the right links yields the tests shown in this commit.

In Listing 14.29, note that we include the assertion

```
assert_not @user.following.empty?
```

which is included to make sure that

```
@user.following.each do |user|
  assert_select "a[href=?]", user_path(user)
end
```
isn’t vacuously true (and similarly for followers). In other words, if

```
@user.following.empty? were true, not a single assert_select would
```

execute in the loop, leading the tests to pass and thereby give us a
false sense of security.

The test suite should now be green:

```
$ rails test
```

Exercises -

1. Verify in a browser that /users/1/followers and /users/1/following
work. Do the image links in the sidebar work as well?

Yes, all urls are working as well as the image links in the sidebar.

2. Comment out the application code needed to turn the assert_select
tests in commit: Tests for following/follower pages red to confirm
they’re testing the right thing.

In the , I've commented out the following line:

```
<%#= link_to gravatar_for(user, size: 30), user %>
```
to make the assert_select turn red.

In addition, we needed to make sure the container div is included with
assert_select parameter because the original code was selecting the
hrefs it the following container:

body > div > div > div > ul.'users follow'

instead of

body > div > div > aside > section.stats > div.user_avatars