Basic access control tests for relationships. red
Now that our views are in order, it’s time to get the follow/unfollow
buttons working. Because following and unfollowing involve creating and
destroying relationships, we need a Relationships controller, which we
generate as usual

As we’ll see in commit: Access control for relationships, enforcing
access control on the Relationships controller actions won’t much
matter, but we’ll still follow our previous practice of enforcing the
security model as early as possible. In particular, we’ll check that
attempts to access actions in the Relationships controller require a
logged-in user (and thus get redirected to the login page), while also
not changing the Relationship count, as shown in this commit.

ERROR["test_create_should_require_logged-in_user", RelationshipsControllerTest, 35.79350590900867]
 test_create_should_require_logged-in_user#RelationshipsControllerTest (35.79s)
AbstractController::ActionNotFound:         AbstractController::ActionNotFound: The action 'create' could not be found for RelationshipsController
            test/controllers/relationships_controller_test.rb:7:in `block (2 levels) in <class:RelationshipsControllerTest>'
            test/controllers/relationships_controller_test.rb:6:in `block in <class:RelationshipsControllerTest>'

=E

ERROR["test_destroy_should_require_logged-in_user", RelationshipsControllerTest, 40.44090314900677]
 test_destroy_should_require_logged-in_user#RelationshipsControllerTest (40.44s)
AbstractController::ActionNotFound:         AbstractController::ActionNotFound: The action 'destroy' could not be found for RelationshipsController
            test/controllers/relationships_controller_test.rb:14:in `block (2 levels) in <class:RelationshipsControllerTest>'
            test/controllers/relationships_controller_test.rb:13:in `block in <class:RelationshipsControllerTest>'

==|
E

Finished in 40.600504s, 1.6749 runs/s, 8.2758 assertions/s.
68 runs, 336 assertions, 0 failures, 2 errors, 0 skips