The Relationships controller.
To get the follow and unfollow buttons to work, all we need to do is
find the user associated with the followed_id in the corresponding form
(i.e., commit: A form for following/unfollowing a user), and then use
the appropriate follow or unfollow method from commit: Utility methods
for following. The full implementation appears in this commit.

We can see from this commit why the security issue mentioned above is
minor: if an unlogged-in user were to hit either action directly (e.g.,
using a command-line tool like curl), current_user would be nil, and in
both cases the action’s second line would raise an exception, resulting
in an error but no harm to the application or its data. It’s best not to
rely on that, though, so we’ve taken the extra step and added an
additional layer of security.

With that, the core follow/unfollow functionality is complete, and any
user can follow or unfollow any other user, as you can verify by
clicking the corresponding buttons in your browser. (We’ll write
integration tests to verify this behavior in Section 14.2.6: Following
tests.) The result of following user #2 is shown in [Figure
14.19](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/unfollowed_user.png)
and [Figure
14.20](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/followed_user.png).

Exercises -

1. Follow and unfollow /users/2 through the web. Did it work?

Yes, the button does work and I see the followers count increase and
decrease on each button click.

2. According to the server log, which templates are rendered in each
case?

In each case the following templates are rendered:

users/show.html.erb within layouts/application:
 layouts/_rails_default.html.erb
 layouts/_shim.html.erb
 layouts/_header.html.erb
 layouts/_footer.html.erb