A form for following/unfollowing a user using Ajax.
A working follow button with Ajax

Although our user following implementation is complete as it stands, we
have one bit of polish left to add before starting work on the status
feed. You may have noticed in Section 14.2.4: A working follow button
the standard way that both the create and destroy actions in the
Relationships controller simply redirect back to the original profile.
In other words, a user starts on another user’s profile page, follows
the other user, and is immediately redirected back to the original page.
It is reasonable to ask why the user needs to leave that page at all.

This is exactly the problem solved by Ajax, which allows web pages to
send requests asynchronously to the server without leaving the page.8
Because adding Ajax to web forms is a common practice, Rails makes Ajax
easy to implement. Indeed, updating the follow/unfollow form partials is
trivial: just change

```
form_for
```

to

```
form_for ..., remote: true
```

and Rails automagically uses Ajax. The updated partials appear in
commit: A form for following/unfollowing a user using Ajax.

The actual HTML generated by this ERb isn’t particularly relevant, but
you might be curious, so here’s a peek at a schematic view (details may
differ):

```
<form action="/relationships/117" class="edit_relationship"
data-remote="true"
      id="edit_relationship_117" method="post">
  .
  .
  .
</form>
```

This sets the variable data-remote="true" inside the form tag, which
tells Rails to allow the form to be handled by JavaScript. By using a
simple HTML property instead of inserting the full JavaScript code (as
in previous versions of Rails), Rails follows the philosophy of
unobtrusive JavaScript.

Having updated the form, we now need to arrange for the Relationships
controller to respond to Ajax requests. We can do this using the
respond_to method, responding appropriately depending on the type of
request. The general pattern looks like this:

```
respond_to do |format|
  format.html { redirect_to user }
  format.js
end
```