Responding to Ajax requests in the Relationships controller.
The syntax is potentially confusing, and it’s important to understand
that in the code above only one of the lines gets executed. (In this
sense, respond_to is more like an if-then-else statement than a series
of sequential lines.) Adapting the Relationships controller to respond
to Ajax involves adding respond_to as above to the create and destroy
actions from commit:The Relationships controller. The result appears as
in this commit. Note the change from the local variable user to the
instance variable @user; in commit:  The Relationships controller there
was no need for an instance variable, but now such a variable is
necessary for use in commit: A form for following/unfollowing a user
using Ajax.