The Ruby JavaScript (RJS) to create/destroy a following relationship.
Listing 14.38: The JavaScript embedded Ruby to create a following
relationship.
app/views/relationships/create.js.erb

Listing 14.39: The Ruby JavaScript (RJS) to destroy a following
relationship.
app/views/relationships/destroy.js.erb

On the other hand, we have yet to respond properly when JavaScript is
enabled. In the case of an Ajax request, Rails automatically calls a
JavaScript embedded Ruby (.js.erb) file with the same name as the
action, i.e., create.js.erb or destroy.js.erb. As you might guess, such
files allow us to mix JavaScript and embedded Ruby to perform actions on
the current page. It is these files that we need to create and edit in
order to update the user profile page upon being followed or unfollowed.

Inside a JS-ERb file, Rails automatically provides the jQuery JavaScript
helpers to manipulate the page using the Document Object Model (DOM).
The jQuery library (which we saw briefly in Section 13.4.2: Image
validation) provides a large number of methods for manipulating the DOM,
but here we will need only two. First, we will need to know about the
dollar-sign syntax to access a DOM element based on its unique CSS id.
For example, to manipulate the follow_form element, we will use the
syntax

```
$("#follow_form")
```

(Recall from commit: A partial for a follow/unfollow form. that this is
a div that wraps the form, not the form itself.) This syntax, inspired
by CSS, uses the # symbol to indicate a CSS id. As you might guess,
jQuery, like CSS, uses a dot . to manipulate CSS classes.

The second method we’ll need is html, which updates the HTML inside the
relevant element with the contents of its argument. For example, to
replace the entire follow form with the string "foobar", we would write

```
$("#follow_form").html("foobar")
```

Unlike plain JavaScript files, JS-ERb files also allow the use of
embedded Ruby, which we apply in the create.js.erb file to update the
follow form with the unfollow partial (which is what should show after a
successful following) and update the follower count. The result is shown
in this commit. This uses the escape_javascript method, which is needed
to escape out the result when inserting HTML in a JavaScript file

Note the presence of line-ending semicolons, which are characteristic of
languages with syntax descended from ALGOL.

The destroy.js.erb file is analogous (Listing 14.39).

With that, you should navigate to a user profile page and verify that
you can follow and unfollow without a page refresh.

Exercises -

1. Unfollow and refollow /users/2 through the web. Did it work?

Yes, it works just as before.

2. According to the server log, which templates are rendered in each
case?

In case of follow:

Rendered users/_unfollow.html.erb (0.0ms)
Rendered relationships/create.js.erb (75.4ms)

In case of unfollow:

Rendering relationships/destroy.js.erb
Rendered users/_follow.html.erb (0.0ms)