A test for the status feed. red
The status feed

We come now to the pinnacle of our sample application: the status feed
of microposts. Appropriately, this section contains some of the most
advanced material in the entire tutorial. The full status feed builds on
the proto-feed from Section 13.3.3: A proto-feed by assembling an array
of the microposts from the users being followed by the current user,
along with the current user’s own microposts. Throughout this section,
we’ll proceed through a series of feed implementations of increasing
sophistication. To accomplish this, we will need some fairly advanced
Rails, Ruby, and even SQL programming techniques.

Because of the heavy lifting ahead, it’s especially important to review
where we’re going. A recap of the final status feed, shown in [Figure
14.5](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/page_flow_home_page_feed_mockup_3rd_edition.png),
appears again in [Figure
14.21](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/page_flow_home_page_feed_mockup_bootstrap.png).

Motivation and strategy

The basic idea behind the feed is simple. [Figure
14.22](https://www.railstutorial.org/book/following_users#fig-user_feed)
shows a sample microposts database table and the resulting feed. The
purpose of a feed is to pull out the microposts whose user ids
correspond to the users being followed by the current user (and the
current user itself), as indicated by the arrows in the diagram.

            |        microposts      |
            | id | content | user_id |            |    user.feed    |
            |------------------------|            |-----------------|
            | 1  | ...     | 1        |---------->| 1 | ...    | 1  |
            | 2  | ...     | 2        |---------->| 2 | ...    | 2  |
            | 3  | ...     | 4        | |-------->| 4 | ...    | 7  |
            | 4  | ...     | 7        |-| |------>| 5 | ...    | 1  |
            | 5  | ...     | 1        |---||----->| 7 | ...    | 8  |
            | 6  | ...     | 18       |    ||---->| 9 | ...    | 10 |
            | 7  | ...     | 8        |----|||--->|10 | ...    | 2  |
            | 8  | ...     | 9        |     ||
            | 9  | ...     | 10       |-----||
            |10  | ...     | 2        |------|

Figure 14.22: The feed for a user (id 1) following users with ids 2, 7,
8, and 10.

Although we don’t yet know how to implement the feed, the tests are
relatively straightforward, so (following the guidelines in Box 3.3)
we’ll write them first. The key is to check all three requirements for
the feed: microposts for both followed users and the user itself should
be included in the feed, but a post from an unfollowed user should not
be included.

As we’ll see in commit: Relationships fixtures for use in
following/follower tests, we’ll be arranging for Michael to follow Lana
but not Archer; based on the fixtures in commit: Adding 30 extra users
to the fixture and commit: Adding a micropost with a different owner,
this means that Michael should see Lana’s posts and his own posts, but
not Archer’s posts. Converting these requirements to assertions and
recalling that the feed is in the User model (commit: A preliminary
implementation for the micropost status feed) gives the updated User
model test shown in this commit.

Of course, the current implementation is just a proto-feed, so the new
test is initially red:

```
$ rails test
```

Exercises

1. Assuming the micropost’s ids are numbered sequentially, with larger
numbers being more recent, what would user.feed.map(&:id) return for the
feed shown in Figure 14.22? Hint: Recall the default scope from Section
13.1.4: Micropost refinements.

user.feed.map(&:id)should return  [2, 10, 8, 7, 2] base on Figure 14.22.