The initial working feed. green
A first feed implementation

With the status feed design requirements captured in the test from
commit: A test for the status feed., we’re ready to start writing the
feed. Since the final feed implementation is rather intricate, we’ll
build up to it by introducing one piece at a time. The first step is to
think of the kind of query we’ll need. We need to select all the
microposts from the microposts table with ids corresponding to the users
being followed by a given user (or the user itself). We might write this
schematically as follows:

```
SELECT * FROM microposts
WHERE user_id IN (<list of ids>) OR user_id = <user id>
```

In writing this code, we’ve guessed that SQL supports an IN keyword that
allows us to test for set inclusion. (Happily, it does.)

Recall from the proto-feed in Section 13.3.3: A proto-feed that Active
Record uses the where method to accomplish the kind of select shown
above, as illustrated in commit: A preliminary implementation for the
micropost status feed.. There, our select was very simple; we just
picked out all the microposts with user id corresponding to the current
user:

```
Micropost.where("user_id = ?", id)
```
Here, we expect it to be more complicated, something like this:

```
Micropost.where("user_id IN (?) OR user_id = ?", following_ids, id)
```

We see from these conditions that we’ll need an array of ids
corresponding to the users being followed. One way to do this is to use
Ruby’s map method, available on any “enumerable” object, i.e., any
object (such as an Array or a Hash) that consists of a collection of
elements.9 We saw an example of this method in Section 4.3.2: Blocks; as
another example, we’ll use map to convert an array of integers to an
array of strings:
```
$ rails console

>> [1, 2, 3, 4].map { |i| i.to_s }
=> ["1", "2", "3", "4"]
```

Situations like the one illustrated above, where the same method gets
called on each element in the collection, are common enough that there’s
a shorthand notation for it (seen briefly in Section 4.3.2: Blocks) that
uses an ampersand & and a symbol corresponding to the method:

```
>> [1, 2, 3, 4].map(&:to_s)
=> ["1", "2", "3", "4"]
```

Using the join method (Section 4.3.1), we can create a string composed
of the ids by joining them on comma-space :

```
>> [1, 2, 3, 4].map(&:to_s).join(', ')
=> "1, 2, 3, 4"
```

We can use the above method to construct the necessary array of followed
user ids by calling id on each element in user.following. For example,
for the first user in the database this array appears as follows:

```
>> User.first.following.map(&:id)
=> [3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
22,
23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51]
```

In fact, because this sort of construction is so useful, Active Record
provides it by default:

```
>> User.first.following_ids
=> [3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51]
```

Here the following_ids method is synthesized by Active Record based on
the has_many :following association (commit: Adding the User model
following association); the result is that we need only append _ids to
the association name to get the ids corresponding to the user.following
collection. A string of followed user ids then appears as follows:

```
>> User.first.following_ids.join(', ')
=> "3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
22,
23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,
41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51"
```

When inserting into an SQL string, though, you don’t need to do this;
the ? interpolation takes care of it for you (and in fact eliminates
some database-dependent incompatibilities). This means we can use
following_ids by itself. As a result, the initial guess of

```
Micropost.where("user_id IN (?) OR user_id = ?", following_ids, id)
````

actually works! The result appears in this commit.

The test suite should be green:

```
$ rails test
```

.=..==.=..==..==...==..==.=.=.=.=.=..==..==...==..==..==.=.=.=.=..==.=.=.=..==..=.=.=.=.=.=.=.=..==..=.==..==..=.=.=.=..==..==..==..=..==.==|
.
Finished in 34.29965s
73 tests, 382 assertions, 0 failures, 0 errors, 0 skips

Finished in 34.733922s, 2.1017 runs/s, 10.9979 assertions/s.
73 runs, 382 assertions, 0 failures, 0 errors, 0 skips

In some applications, this initial implementation might be good enough
for most practical purposes, but this commit isn’t the final
implementation; see if you can make a guess about why not before moving
on to the next section. (Hint: What if a user is following 5000 other
users?)

Exercises -

1. In this commit, remove the part of the query that finds the user’s
own posts. Which test in commit: A test for the status feed breaks?

```
def feed
  Micropost.where("user_id IN (?)", id)
end
```

1. Error

test/integration/microposts_interface_test.rb:12:
test_micropost_interface

Expected at least 1 element matching "div.pagination", found 0..
        Expected 0 to be >= 1.
2. Error

test/models/user_test.rb:111:in `block (2 levels) in <class:UserTest>'
test/models/user_test.rb:110:in `block in <class:UserTest>'

test_feed_should_have_the_right_posts

 Expected false to be truthy.

2. In this commit, remove the part of the query that finds the followed
users’ posts. Which test in commit: A test for the status feed breaks?

```
def feed
  Micropost.where("user_id = ?", id)
end
```
FAIL["test_feed_should_have_the_right_posts", UserTest,
20.52635398399434]
 test_feed_should_have_the_right_posts#UserTest (20.53s)
        Expected false to be truthy.
        test/models/user_test.rb:106:in `block (2 levels) in
<class:UserTest>'
        test/models/user_test.rb:105:in `block in <class:UserTest>'

3. How could you change the query in this commit to have the feed
erroneously return microposts of unfollowed users, thereby breaking the
third test in commit: A test for the status feed? Hint: Returning all
the microposts would do the trick.

```
def feed
  Micropost.all
end
```

FAIL["test_feed_should_have_the_right_posts", UserTest,
21.64591703301994]
test_feed_should_have_the_right_posts#UserTest (21.65s)
       Expected true to be nil or false
       test/models/user_test.rb:116:in `block (2 levels) in
<class:UserTest>'
       test/models/user_test.rb:115:in `block in <class:UserTest>'