Using key-value pairs in the feed’s where method. green
As hinted at in the last section, the feed implementation in Section 14.3.2 doesn’t scale well when the number of microposts in the feed is large, as would likely happen if a user were following, say, 5000 other users. In this section, we’ll reimplement the status feed in a way that scales better with the number of followed users.

The problem with the code in Section 14.3.2: A first feed implementation is that following_ids pulls all the followed users’ ids into memory, and creates an array the full length of the followed users array. Since the condition in commit: A test for the status feed actually just checks inclusion in a set, there must be a more efficient way to do this, and indeed SQL is optimized for just such set operations. The solution involves pushing the finding of followed user ids into the database using a subselect.

We’ll start by refactoring the feed with the slightly modified code in this commit.

As preparation for the next step, we have replaced

```
Micropost.where("user_id IN (?) OR user_id = ?", following_ids, id)
```

with the equivalent

```
Micropost.where("user_id IN (:following_ids) OR user_id = :user_id",
                following_ids: following_ids, user_id: id)
```

The question mark syntax is fine, but when we want the same variable inserted in more than one place, the second syntax is more convenient.

The above discussion implies that we will be adding a second occurrence of user_id in the SQL query. In particular, we can replace the Ruby code

```
following_ids
```
with the SQL snippet

```
following_ids = "SELECT followed_id FROM relationships
                 WHERE  follower_id = :user_id"
```
This code contains an SQL subselect, and internally the entire select for user 1 would look something like this:

```
SELECT * FROM microposts
WHERE user_id IN (SELECT followed_id FROM relationships
                  WHERE  follower_id = 1) OR user_id = 1
```

This subselect arranges for all the set logic to be pushed into the database, which is more efficient.