The final implementation of the feed. green
With this foundation, we are ready for a more efficient feed
implementation, as seen in commit: The final implementation of the feed.
Note that, because it is now raw SQL, the following_ids string is
interpolated, not escaped.

This code has a formidable combination of Rails, Ruby, and SQL, but it
does the job, and does it well:

```
$ rails test
```

.=.=.=.=.=.=..==..==..=.=.=.=.=.=.=.=.=..==..==..=.=.=.=..==.=.=..==.=.=.=..==..=.=.=.=.=.=.=.=...==..===..==..=.=.=.=.=.=.=.=..==..=.=.=.==|
.
Finished in 33.37810s
73 tests, 382 assertions, 0 failures, 0 errors, 0 skips

Finished in 33.614067s, 2.1717 runs/s, 11.3643 assertions/s.
73 runs, 382 assertions, 0 failures, 0 errors, 0 skips

Of course, even the sub-select won’t scale forever. For bigger sites,
you would probably need to generate the feed asynchronously using a
background job, but such scaling subtleties are beyond the scope of this
tutorial.

With the code in this commit, our status feed is now complete. Recall
from Section 13.3.3: A proto-feed that the Home page already includes
the feed. In Chapter 13: User microposts, the result was only a
proto-feed ([Figure
13.14](https://softcover.s3.amazonaws.com/636/ruby_on_rails_tutorial_4th_edition/images/figures/home_with_proto_feed_3rd_edition.png)),
but with the implementation in commit: The final implementation of the
feed as seen in [Figure
14.23](https://www.railstutorial.org/book/following_users#fig-home_page_with_feed)
the Home page now shows the full feed.