Testing the feed HTML. green
1. Write an integration test to verify that the first page of the feed
appears on the Home page as required. A template appears in this commit.

Done, its located in /test/integration/following_test.rb.

.=.=.=.=.=.=..==..=.=.=.=.=.=..==.=.=.=...==...===...===..==..=.=.=.==..=.=.=..==..=.==..=.=.=.=.=.=...===..==..==...==.=..==.=...==.=..==.==|
.
Finished in 26.00853s
74 tests, 442 assertions, 0 failures, 0 errors, 0 skips

Finished in 26.236662s, 2.8205 runs/s, 16.8467 assertions/s.
74 runs, 442 assertions, 0 failures, 0 errors, 0 skips

2. Note that in this commit, escapes the expected HTML using
CGI.escapeHTML (which is closely related to the CGI.escape method we
used in Section 11.2.3: Email tests to escape URLs). Why is escaping the
HTML necessary in this case? Hint: Try removing the escaping and
carefully inspect the page source for the micropost content that doesn’t
match. Using the search feature of your terminal shell (Cmd-F on Ctrl-F
on most systems) to find the word “sorry” may prove particularly
helpful.

Here is what would in the test would look like:

```
test "feed on Home page" do
  get root_path
  @user.feed.paginate(page: 1).each do |micropost|
    assert_match micropost.content, response.body
  end
end
```

It is necessary use CGI.escapeHTML because it escape special characters
in HTML, namely &"<>. When this not applied it hard for assert_match try
to expected to match micropost.content with the response.body HTML.